Move config builder's UI to the demo app
