+++ demo-app/src/pages/ot/config-builder/FrameWrapper.tsx
@@ -1,0 +1,67 @@
+import {
+  useConfigBuilderContext,
+  type FrameWrapper,
+} from '@open-truss/open-truss'
+
+const ConfigBuilderFrameWrapper: FrameWrapper = ({
+  children,
+  configPath,
+  frame,
+}) => {
+  const { framesPath, setFramesPath, deleteFrame } = useConfigBuilderContext()
+
+  let nestingHere = false
+  // When we add a frame we set its `.frames` to an empty array if
+  // it supports children, so here we just check if that field is truthy.
+  const canNestFrames = Boolean(frame?.frames)
+  if (canNestFrames) {
+    // `framesPath` is the current nesting component's `.frames` so to ask
+    // if this component is the nester, we remove `.frames` from the end.
+    nestingHere = configPath === framesPath.replace(/\.frames$/, '')
+  }
+
+  const nestHere = (e: React.MouseEvent): void => {
+    e.stopPropagation()
+    if (nestingHere) {
+      // We're toggling off frame nesting.
+      setFramesPath(null)
+      return
+    }
+    // Set framesPath to this component's frames array.
+    setFramesPath(`${configPath}.frames`)
+  }
+
+  const deleteThisFrame = (e: React.MouseEvent): void => {
+    e.stopPropagation()
+    deleteFrame(configPath)
+  }
+
+  return (
+    <div
+      style={{
+        position: 'relative',
+        border: nestingHere ? '1px solid pink' : undefined,
+      }}
+    >
+      <div
+        style={{
+          position: 'absolute',
+          right: 0,
+          top: 0,
+          display: 'flex',
+          padding: '5px',
+        }}
+      >
+        {canNestFrames && (
+          <button onClick={nestHere}>
+            {nestingHere ? 'Stop nesting' : 'Nest here'}
+          </button>
+        )}
+        <button onClick={deleteThisFrame}>Delete</button>
+      </div>
+      {children}
+    </div>
+  )
+}
+
+export default ConfigBuilderFrameWrapper
