+++ packages/open-truss/src/pages.ts
@@ -1,2 +1,1 @@
 export * from './pages/playground-page'
-export * from './pages/config-builder-page/index'
